Refactor HTTP proxy into separate file
